Add dark background support to Solarized
The Solarized colortheme now adapts to the value of the 'background'
option. If 'background' is set to 'light', nothing changes, but when it
is set to 'dark', the 'base0{0,1,2,3}' and the 'base{0,1,2,3}' colors
are swapped, as prescribed by http://ethanschoonover.com/solarized .

The values of the base colors are hardcodded twice, since using variables did
not seem to work correctly.